remove tailwind + add stitches

+++ app/routes/polls/$pollId.tsx
@@ -4,19 +4,27 @@ import { Form, useCatch, useLoaderData } from "@remix-run/react";
 import invariant from "tiny-invariant";
 import type { Poll, Answer } from "~/models/poll.server";
 import { getAnswers } from "~/models/poll.server";
-import { getPoll, deletePoll } from "~/models/poll.server";
+import { getPoll, deletePoll, publishPoll } from "~/models/poll.server";
 import { requireUserId } from "~/session.server";
-import { useOptionalUser } from "~/utils";
+import { useUser } from "~/utils";
 
 type LoaderData = {
   poll: Poll;
   answers: Answer[];
 };
 
+type ActionData = {
+  errors?: {
+    id?: string;
+    publish?: string;
+  };
+};
+
 export const loader: LoaderFunction = async ({ request, params }) => {
+  const userId = await requireUserId(request);
   invariant(params.pollId, "pollId not found");
 
-  const poll = await getPoll({ id: params.pollId });
+  const poll = await getPoll({ id: params.pollId, userId });
   if (!poll) {
     throw new Response("Not Found", { status: 404 });
   }
@@ -29,30 +37,69 @@ export const action: ActionFunction = async ({ request, params }) => {
   const userId = await requireUserId(request);
   invariant(params.pollId, "pollId not found");
 
-  await deletePoll({ userId, id: params.pollId });
+  const data = await request.formData();
+  const action = data.get("_action");
 
-  return redirect("/polls");
+  if (typeof action !== "string" || action.length === 0) {
+    return json<ActionData>(
+      { errors: { id: "Action is required" } },
+      { status: 400 }
+    );
+  }
+
+  if (action === "delete") {
+    await deletePoll({ userId, id: params.pollId });
+    return redirect("/polls");
+  }
+
+  if (action === "publish") {
+    const publishInputValue = data.get("publish");
+    if (
+      typeof publishInputValue !== "string" ||
+      !["true", "false"].includes(publishInputValue)
+    ) {
+      return json<ActionData>(
+        { errors: { publish: "Publish input value is required" } },
+        { status: 400 }
+      );
+    }
+    const isPublished = publishInputValue === "true" ? true : false;
+    await publishPoll({ userId, id: params.pollId, newState: !isPublished });
+    return redirect("/polls/" + params.pollId);
+  }
+
+  return null;
 };
 
 export default function PollDetailsPage() {
   const data = useLoaderData() as LoaderData;
-  const user = useOptionalUser();
+  // const transition = useTransition();
+  const user = useUser();
 
   return (
     <div>
-      <h3 className="text-2xl font-bold">{data.poll.title}</h3>
-      <p className="py-6">{data.poll.body}</p>
-      <hr className="my-4" />
+      <h3>{data.poll.title}</h3>
+      <p>{data.poll.body}</p>
+      <hr />
       {data.answers.map((answer) => (
-        <p key={answer.id}>{answer.title}</p>
+        <p key={answer.id}>
+          {answer.title} - {answer.count}
+        </p>
       ))}
-      <hr className="my-4" />
+      <hr />
+      <Form method="post">
+        <input
+          type="hidden"
+          name="publish"
+          value={data.poll.publish ? "true" : "false"}
+        />
+        <button type="submit" name="_action" value="publish">
+          Publish : {data.poll.publish ? "Yes" : "No"}
+        </button>
+      </Form>
       {user && (
         <Form method="post">
-          <button
-            type="submit"
-            className="rounded bg-blue-500  py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400"
-          >
+          <button type="submit" name="_action" value="delete">
             Delete
           </button>
         </Form>
@@ -71,7 +118,7 @@ export function CatchBoundary() {
   const caught = useCatch();
 
   if (caught.status === 404) {
-    return <div>Note not found</div>;
+    return <div>Poll not found</div>;
   }
 
   throw new Error(`Unexpected caught response with status: ${caught.status}`);
